fix(remount): resolve race condition in remount code (#122)
When the remount feature is enabled process will be monitoring for mount points on that node. If mount points are marked as RO, process will launch go routines for them in a loop. Since these go routine are using the same stack variable vol when the loop iterates, change in value of the variable will lead to volumes not getting removed from transition list.

Signed-off-by: Payes Anand <payes.anand@mayadata.io>